backend/vision/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
opencv-python-headless==4.8.1.78
Pillow==10.1.0
ImageHash==4.3.1
numpy==1.26.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Optional heavy ML stack (enable by setting ENABLE_CLIP=true and uncommenting below)
# torch==2.1.1 ; platform_system != 'Windows' or platform_machine == 'x86_64'
# torchvision==0.16.1
# git+https://github.com/openai/CLIP.git@main#egg=clip

# Notes:
# - Dev/test dependencies removed to keep runtime image lean.
# - Pin versions for reproducible Docker caching.
